Change 'Type of column'

+++ database/migrations/2020_09_25_053310_create_post_tags_table.php
@@ -15,8 +15,8 @@ class CreatePostTagsTable extends Migration
     {
         Schema::create('post_tags', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('tag_id');
-            $table->unsignedInteger('post_id');
+            $table->unsignedBigInteger('tag_id');
+            $table->unsignedBigInteger('post_id');
             $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
             $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
             $table->timestamps();
